Add back conftest
